Improve mailcheck flexibility
Now, when the persistent process is started, it always uses the quiet
interface.

When the non-persistent process is started, it passes along any extra
arguments that the script was started with. Thus, if you only want to do
a single quick sync, you could do something like:

    $ mailcheck.sh -q

If you want to do a full sync, you can still do a simple:

    $ mailcheck.sh